Rework example client side
With introduction of refresh it becomes tedious to reimplement the
client side for every single frontend. Thus, rework it to be a generic
module that can be reused. Don't want to pull in a full dependency of
another OAuth client implementation still which may come with their own
peculiarities and feature divergence.