Update rocket example for refreshing
While this no longer showcases signed tokens, that implementation detail
is not as valuable as clarifying the mechanism for refresh tokens. It
remains open whether some unified token flow for both token types
(access and refresh) is more relevant in the real world and thus need
specific support. If so, this should be incorporated into the example as
well.